update chapter and user routes
