+++ server/src/config/configPassport.ts
@@ -8,8 +8,9 @@ import {
 
 interface IUserWithPassword {
   _id: string;
-  firstName: string;
-  lastName: string;
+  city: string;
+  state: string;
+  isAcceptingRequests: boolean;
   email: string;
   password?: string;
   verified: boolean;
