+++ server/src/routes/auth.route.ts
@@ -6,26 +6,15 @@ import express from 'express';
 import {
   login,
   logout,
-  register,
   approve,
   sendResetPasswordEmail,
   resetPassword,
   verifyAccount,
-  registerInvite,
 } from '../controllers/auth.controller.ts';
 import { isAuthenticated } from '../controllers/auth.middleware.ts';
 import 'dotenv/config';
 
 const router = express.Router();
-
-/**
- * A POST route to register a user. Expects a JSON body with the following fields:
- * - firstName (string) - The first name of the user
- * - lastName (string) - The last name of the user
- * - email (string) - The email of the user
- * - password (string) - The password of the user
- */
-router.post('/register', register);
 
 /**
  * A POST route to verify a user's account. Expects a JSON body with the following fields:
@@ -66,11 +55,5 @@ router.post('/reset-password', resetPassword);
  * and 401 unauthorized if the user is not authenticated.
  */
 router.get('/authstatus', isAuthenticated, approve);
-
-/**
- * A POST register a user from an invite. If the information and invite are valid
- * a new account is created. Otherwise a 400 bad request error is returned
- */
-router.post('/register-invite', registerInvite);
 
 export default router;
